refactor(web): move api path
