+++ web/__tests__/server/api/courses/index.get.test.ts
@@ -1,47 +1,50 @@
-import { testSongList } from '@ddradar/core/__tests__/data'
+import { testCourseData } from '@ddradar/core/__tests__/data'
 import { fetchList } from '@ddradar/db'
 import { useQuery } from 'h3'
 import { beforeEach, describe, expect, test, vi } from 'vitest'
 
-import searchSongs from '~/server/api/v1/songs.get'
-
-import { createEvent } from '../test-util'
+import { createEvent } from '~/__tests__/server/test-util'
+import searchCourses from '~/server/api/v1/courses/index.get'
 
 vi.mock('@ddradar/db')
 vi.mock('h3')
 
-describe('GET /api/v1/songs', () => {
+describe('GET /api/v1/courses', () => {
   beforeEach(() => {
     vi.mocked(fetchList).mockClear()
   })
 
-  const defaultCond = { condition: 'c.nameIndex >= 0' }
+  const defaultCond = { condition: 'c.nameIndex < 0' }
   test.each([
     [undefined, undefined, []],
     ['-1', '-1', []],
-    ['0.5', '0.5', []],
     ['100', '100', []],
-    ['25', undefined, [{ condition: 'c.nameIndex = @', value: 25 }]],
-    [undefined, '10', [{ condition: 'c.series = @', value: 'DDR X' }]],
+    ['0.5', '0.5', []],
+    ['1', undefined, [{ condition: 'c.nameIndex = @', value: -1 }]],
     [
-      ['25'],
-      '0',
+      undefined,
+      '17',
+      [{ condition: 'c.series = @', value: 'DanceDanceRevolution A20 PLUS' }],
+    ],
+    [
+      '2',
+      '18',
       [
-        { condition: 'c.nameIndex = @', value: 25 },
-        { condition: 'c.series = @', value: 'DDR 1st' },
+        { condition: 'c.nameIndex = @', value: -2 },
+        { condition: 'c.series = @', value: 'DanceDanceRevolution A3' },
       ],
     ],
   ])(
     '?name=%s&series=%s calls fetchList(..., ..., %o)',
-    async (name, series, expected) => {
+    async (type, series, expected) => {
       // Arrange
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      vi.mocked(fetchList).mockResolvedValue([...testSongList] as any)
-      vi.mocked(useQuery).mockReturnValue({ name, series })
+      vi.mocked(fetchList).mockResolvedValue([testCourseData] as any)
+      vi.mocked(useQuery).mockReturnValue({ type, series })
       const event = createEvent()
 
       // Act
-      const songs = await searchSongs(event)
+      const songs = await searchCourses(event)
 
       // Assert
       expect(event.res.statusCode).toBe(200)
@@ -57,7 +60,7 @@ describe('GET /api/v1/songs', () => {
     const event = createEvent()
 
     // Act
-    const songs = await searchSongs(event)
+    const songs = await searchCourses(event)
 
     // Assert
     expect(event.res.statusCode).toBe(404)
