+++ web/server/api/v1/courses/index.get.ts
